[svn-to-darcs @ 14] Added a test document.
Added a few more to-do items.

Fixed ignores to include .dvi and .ps.

All sorts of revisions to ocamltex.dtx. Running "latex ocamltex.ins" should now
reconstitute approximately the same files as the 0.5 release.  Testing still
needs to be done, and there is a lot of explanation of ocamltex.ml in
ocamltex.dtx that is still needed.